relocate file to src
