+++ src/components/FriendsActivity.tsx
